+++ R/gghistostats.R
@@ -28,7 +28,12 @@
 #'   `FALSE`, no statistical tests will be run.
 #' @param centrality.para Decides *which* measure of central tendency ("mean" or
 #'   "median") is to be displayed as a vertical line.
-#' @param centrality.colour Decides colour for the vertical line.
+#' @param centrality.colour Decides colour for the vertical line for centrality
+#'   parameter (Default: `"blue"`).
+#' @param test.value.line Decides whether test value is to be displayed as a
+#'   vertical line (Default: `FALSE`).
+#' @param test.value.colour Decides colour for the vertical line denoting test
+#'   value (Default: `"black"`).
 #' @param binwidth.adjust If set to `TRUE`, you can use it to pick better value
 #'   with the `binwidth` argument to `stat_bin()`.
 #' @param binwidth The width of the bins. Can be specified as a numeric value,
@@ -73,6 +78,7 @@
 #' bf.prior = 0.8,
 #' test.value = 3,
 #' centrality.para = "mean",
+#' test.value.line = TRUE,
 #' binwidth.adjust = TRUE,
 #' binwidth = 0.10
 #' )
@@ -123,22 +129,24 @@ utils::globalVariables(
 # function body
 gghistostats <-
   function(data = NULL,
-             x,
-             xlab = NULL,
-             title = NULL,
-             subtitle = NULL,
-             caption = NULL,
-             type = "parametric",
-             test.value = 0,
-             bf.prior = 0.707,
-             bf.message = TRUE,
-             k = 3,
-             results.subtitle = TRUE,
-             centrality.para = NULL,
-             centrality.colour = "blue",
-             binwidth.adjust = FALSE,
-             binwidth = NULL,
-             messages = TRUE) {
+           x,
+           xlab = NULL,
+           title = NULL,
+           subtitle = NULL,
+           caption = NULL,
+           type = "parametric",
+           test.value = 0,
+           bf.prior = 0.707,
+           bf.message = TRUE,
+           k = 3,
+           results.subtitle = TRUE,
+           centrality.para = NULL,
+           centrality.colour = "blue",
+           test.value.line = FALSE,
+           test.value.colour = "black",
+           binwidth.adjust = FALSE,
+           binwidth = NULL,
+           messages = TRUE) {
     # if data is not available then don't display any messages
     if (is.null(data)) {
       messages <- FALSE
@@ -149,20 +157,16 @@ gghistostats <-
     # preparing a dataframe out of provided inputs
     if (!is.null(data)) {
       # preparing labels from given dataframe
-      lab.df <- colnames(dplyr::select(
-        .data = data,
-        !!rlang::enquo(x)
-      ))
+      lab.df <- colnames(dplyr::select(.data = data,
+                                       !!rlang::enquo(x)))
       # if xlab is not provided, use the variable x name
       if (is.null(xlab)) {
         xlab <- lab.df[1]
       }
       # if dataframe is provided
       data <-
-        dplyr::select(
-          .data = data,
-          x = !!rlang::enquo(x)
-        )
+        dplyr::select(.data = data,
+                      x = !!rlang::enquo(x))
     } else {
       # if vectors are provided
       data <-
@@ -340,24 +344,19 @@ gghistostats <-
 
     # if the user wants to adjust the binwidth
     if (isTRUE(binwidth.adjust)) {
-      plot <- ggplot2::ggplot(
-        data = data,
-        mapping = ggplot2::aes(x = x)
-      ) +
+      plot <- ggplot2::ggplot(data = data,
+                              mapping = ggplot2::aes(x = x)) +
         ggplot2::stat_bin(
           col = "black",
           alpha = 0.7,
           binwidth = binwidth,
           na.rm = TRUE,
-          mapping = ggplot2::aes(
-            y = ..count..,
-            fill = ..count..
-          )
+          mapping = ggplot2::aes(y = ..count..,
+                                 fill = ..count..)
         ) +
         ggplot2::scale_fill_gradient("count",
-          low = "green",
-          high = "red"
-        ) +
+                                     low = "green",
+                                     high = "red") +
         ggstatsplot::theme_mprl() +
         ggplot2::labs(
           x = xlab,
@@ -367,10 +366,8 @@ gghistostats <-
         )
     } else {
       # if not, use the defaults
-      plot <- ggplot2::ggplot(
-        data = data,
-        mapping = ggplot2::aes(x = x)
-      ) +
+      plot <- ggplot2::ggplot(data = data,
+                              mapping = ggplot2::aes(x = x)) +
         ggplot2::geom_histogram(
           col = "black",
           alpha = 0.7,
@@ -378,9 +375,8 @@ gghistostats <-
           na.rm = TRUE
         ) +
         ggplot2::scale_fill_gradient("count",
-          low = "green",
-          high = "red"
-        ) +
+                                     low = "green",
+                                     high = "red") +
         ggstatsplot::theme_mprl() +
         ggplot2::labs(
           x = xlab,
@@ -435,6 +431,17 @@ gghistostats <-
       }
     }
 
+    # if central tendency parameter is to be added
+    if (isTRUE(test.value.line)) {
+      plot <- plot +
+        ggplot2::geom_vline(
+          xintercept = test.value,
+          linetype = "dashed",
+          color = test.value.colour,
+          size = 1.2,
+          na.rm = TRUE
+        )
+    }
     # if caption is provided then use combine_plots function later on to add this caption
     # add caption with bayes factor
     if (isTRUE(results.subtitle)) {
@@ -444,8 +451,7 @@ gghistostats <-
             if (!is.null(bf.caption)) {
               plot <-
                 ggstatsplot::combine_plots(plot,
-                  caption.text = bf.caption.text
-                )
+                                           caption.text = bf.caption.text)
             }
           }
         }
@@ -470,11 +476,9 @@ gghistostats <-
             ": p-value = "
           ),
           crayon::yellow(
-            ggstatsplot::specify_decimal_p(
-              x = ad_norm$p.value[[1]],
-              k,
-              p.value = TRUE
-            )
+            ggstatsplot::specify_decimal_p(x = ad_norm$p.value[[1]],
+                                           k,
+                                           p.value = TRUE)
           )
         ))
       }